requirements.txt
agno
litellm
google-api-python-client
python-telegram-bot
pyyaml
Pillow